added Col instead of div
