+++ src/assets/scss/custome.scss
@@ -4,11 +4,5 @@
 
 
 @import "styles/navbar";
-<<<<<<< HEAD:my-app/src/assets/scss/custome.scss
 @import "styles/home";
 @import "styles/signupPage"
-=======
-
-@import "styles/home";
-
->>>>>>> 86584da4fcc9fb5a37d32581b6a7bcb3407602ab:src/assets/scss/custome.scss
