+++ src/assets/scss/styles/_signupPage.scss
@@ -19,3 +19,34 @@
 .signup-form {
     margin-top: 30px;
 }
+
+.signup-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+@media (min-width:320px) {
+    .signup-button {
+        padding-right: 10px;
+    }
+}
+
+@media (min-width:768px) {
+    .signup-button {
+        margin-right:199px;
+    }
+}
+
+@media (min-width:1024px) {
+    .signup-button {
+        margin-right: 273px;
+    }
+}
+
+@media (min-width:1440px) {
+    .signup-button {
+        margin-right:373px;
+    }
+}
