+++ src/constants/index.js
@@ -5,21 +5,10 @@ export const required = value => (value ? undefined : 'Required')
 
 
 
-export const date = value => (value ? undefined : 'Required')
-
-export const checkBox = value => (value ? undefined : "Required")
-
-export const email = value =>
+export const emailValidate = value =>
     value && !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(value)
         ? 'Invalid email address'
         : null
-
-
-export const select = value => (value ? undefined : "Required")
-
-
-
-
 
 const maxLength = max => value =>
     value && value.length > max ? `Must be ${max} characters or less` : null
