+++ src/redux/sagas/index.js
@@ -1,1 +1,1 @@
-import { all } from 'redux-saga/effects'
+// import { all } from 'redux-saga/effects'
